Move current RBAC to a namespaced Filterer class
